# IGNITE-24045 Sql. Replace UUID.randomUUID() with custom generator for query id