https://issues.apache.org/jira/browse/IGNITE-24045

Result of the benchmark on my laptop (MBP M3) as follow:

```
Before
Benchmark                         (clusterSize)  (fsync)  Mode  Cnt   Score   Error  Units
SelectBenchmark.sqlGet_1_thread               1    false  avgt   20   6.061 ± 0.137  us/op
SelectBenchmark.sqlGet_4_threads              1    false  avgt   20   8.773 ± 0.848  us/op
SelectBenchmark.sqlGet_8_threads              1    false  avgt   20  16.954 ± 1.540  us/op

After
Benchmark                         (clusterSize)  (fsync)  Mode  Cnt   Score   Error  Units
SelectBenchmark.sqlGet_1_thread               1    false  avgt   20   5.758 ± 0.099  us/op
SelectBenchmark.sqlGet_4_threads              1    false  avgt   20   8.473 ± 0.958  us/op
SelectBenchmark.sqlGet_8_threads              1    false  avgt   20  15.295 ± 1.334  us/op
```

------------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)